PS-1956: Adjust datatype for some timers used for extended slow query log
The ulong and ulint are used for such timers as innodb_io_reads_wait_timer,
innodb_lock_que_wait_timer and innodb_innodb_que_wait_timer. They are
32-bit at some platforms which may be not enough in some cases.
Adjust corresponding datatypes to 64-bit.

+++ include/mysql/plugin.h
@@ -812,8 +812,7 @@ enum mysql_trx_stat_type {
   @param[in]    value   the value of statistics
 */
 void thd_report_innodb_stat(MYSQL_THD thd, unsigned long long trx_id,
-                            enum mysql_trx_stat_type type,
-                            unsigned long long value);
+                            enum mysql_trx_stat_type type, uint64_t value);
 
 unsigned long thd_log_slow_verbosity(const MYSQL_THD thd);
